requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
discord.py==1.7.2
Django==3.0.5
django-allauth==0.44.0
django-dbbackup==3.3.0
django-environ==0.4.5
django-rename-app==0.1.3
djongo==1.3.6
dnspython==2.1.0
environ==1.0
idna==2.10
multidict==5.1.0
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
pymongo==3.11.4
PyNaCl==1.4.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.2.4
typing-extensions==3.10.0.0
urllib3==1.26.4
wrapt==1.12.1
yarl==1.6.3
